slfhelper: Useful functions for working with the Source Linkage Files

This package provides helper functions for working with the Source Linkage Files (SLFs). The functions are mainly focused on making the first steps of analysis easier. They can read and filter the files efficiently using minimal code.

See also

Author

Maintainer: Megan McNicol megan.mcnicol2@phs.scot

Authors:

James McMahon james.mcmahon@phs.scot (ORCID)

Other contributors:

Public Health Scotland phs.source@phs.scot [copyright holder]